Add OrderStatus field to Order class
Enum file "OrderStatus" has been added for maintaining the status of an order in a more organized manner. The status can be "Processing", "In Delivery" or "Completed". Similarly, the Order class has been modified to include a new field "OrderStatus". This would help in better tracking and handling of orders.